+++ client/src/pages/lab_management/statistics/StatisticsPage.tsx
@@ -4,12 +4,9 @@ import { Alert, Box, Button, CardContent, CardHeader, FormControl, FormGroup, Gr
 import CountCard from "./CountCard";
 import { gql, useLazyQuery, useQuery } from "@apollo/client";
 import AdminPage from "../../AdminPage";
-import { Card } from "@material-ui/core";
-import { BarChart, Gauge, LineChart } from "@mui/x-charts";
 import RequestWrapper from "../../../common/RequestWrapper";
-import TimeAgo from 'react-timeago'
-import GET_ROOMS from "../../../queries/getRooms";
-import { GET_EQUIPMENT_SESSIONS, GET_MODULE_SCORES, GET_NUM_EQUIPMENT_SESSIONS_TODAY, GET_NUM_NEW_USERS, GET_NUM_ROOM_SWIPES_TODAY, GET_NUM_SITE_VISITS, GET_ROOM_SWIPE_COUNTS, GET_ZONE_HOURS } from "./statisticsQueries";
+import { GET_NUM_EQUIPMENT_SESSIONS_TODAY, GET_NUM_NEW_USERS, GET_NUM_ROOM_SWIPES_TODAY, GET_NUM_SITE_VISITS, GET_ROOM_SWIPE_COUNTS, GET_ZONE_HOURS } from "./statisticsQueries";
+import { EquipmentStats } from "./equipment/EquipmentStats";
 
 
 function getMonthToPresentBounds(): { startOfMonth: Date, today: Date } {
@@ -65,7 +62,7 @@ export default function StatisticsPage() {
   }
 
   return (
-    <AdminPage title="Statistics" maxWidth="1250px">
+    <AdminPage title="Statistics">
       <Box>
         <Typography variant="h4">Today's Numbers</Typography>
         <Stack direction={"row"} flexWrap={"wrap"}>
@@ -85,9 +82,7 @@ export default function StatisticsPage() {
         <Typography variant="body2">* Only counts ACS-connected equipment</Typography>
       </Box>
 
-      <Box mt={5}>
-
-      </Box>
+      <EquipmentStats />
 
     </AdminPage>
   );
